backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
python-docx==0.8.11
reportlab==3.6.12
werkzeug==2.2.3
markupsafe==2.1.3
Jinja2==3.1.2
